requirements.txt
fastapi
uvicorn
billboard.py
